Changed const Strings& ref inside NetFiler to not be a reference.
The user can't know that NefFilter is storing a reference to the
parameter passed at construction. Documenting the behaviour is not the
right solution to keep the reference beacuse it'd be easy to overlook.
Passing a pointer would be a better choice because it makes more explicit
the fact that the class is going to store the reference (at least makes the
user wonder what would happen to the pointer).
To not change the API a copy will be done instead.